third_party/blink/web_tests/editing/selection/keep-selection-after-set-focus-expected.txt

Selection should be kept after focus changed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS selection.anchorNode is sample
PASS selection.anchorOffset is 3
PASS selection.focusNode is sample
PASS selection.focusOffset is 6
PASS successfullyParsed is true

TEST COMPLETE
